added code to schwierigkeitsgrads

+++ db/migrate/20140811080517_add_zeitdauer_to_wanderungens.rb
@@ -1,6 +1,6 @@
 class AddZeitdauerToWanderungens < ActiveRecord::Migration
   def change
-    add_column :wanderungens, :zeitdauer, :integer
+    add_column :wanderungens, :zeitdauer, :float
     add_index :wanderungens, :zeitdauer
   end
 end
